requirements.txt
google-cloud-vision
python-docx
PyPDF2
Pillow
opencv-python
pytesseract
```